demand-forecasting-platform/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
joblib==1.4.2
pydantic==2.8.2
pytest==8.3.2
httpx==0.27.0
mlflow==2.16.2
pyyaml==6.0.2